Added rating strings and current phase
As cadets promote, they advance into different phases, which are scored
differently in inspections. To prevent overall inspection scores from
changing, the inspection card will store the current phase of the cadet
so that if he/she promotes to the next phase, the overall score remains
unchanged.

Oh, and rating strings meaning conversion from the enum to a human-
readable string